move 'maxBy' tests to separate file
