+++ types/ramda/ramda-tests.ts
@@ -1318,24 +1318,6 @@ type Pair = KeyValuePair<string, number>;
 };
 
 () => {
-    function cmp(obj: { x: R.Ord }) {
-        return obj.x;
-    }
-
-    const a = {x: 1};
-    const b = {x: 2};
-    const c = {x: 3};
-    const d = {x: "a"};
-    const e = {x: "z"};
-    const f = {x: new Date(0)};
-    const g = {x: new Date(60 * 1000)};
-    R.maxBy(cmp, a, c); // => {x: 3}
-    R.maxBy(cmp)(a, c); // => {x: 3}
-    R.maxBy(cmp)(a)(b);
-    R.maxBy(cmp)(f)(g);
-};
-
-() => {
     const a: number = R.mean([2, 7, 9]); // => 6
     const b: number = R.mean([]); // => NaN
 };
