+++ types/ramda/tsconfig.json
@@ -130,6 +130,7 @@
         "test/mapAccumRight-tests.ts",
         "test/mapObjIndexed-tests.ts",
         "test/mathMod-tests.ts",
-        "test/max-tests.ts"
+        "test/max-tests.ts",
+        "test/maxBy-tests.ts"
     ]
 }
